Common: Attempt to handle files that are out of iso range
