+++ common/iso_path_reader/methods/compressed.py
@@ -2,9 +2,10 @@ import datetime
 import io
 
 from common.iso_path_reader.methods.base import IsoPathReader
+from common.iso_path_reader.methods.chunked_hash_trait import ChunkedHashTrait
 
 
-class CompressedPathReader(IsoPathReader):
+class CompressedPathReader(ChunkedHashTrait, IsoPathReader):
     def __init__(self, iso, fp):
         super().__init__(iso, fp)
         self.files = {}
@@ -40,13 +41,6 @@ class CompressedPathReader(IsoPathReader):
 
     def get_file_size(self, file):
         return file.file_size
-
-    def get_file_hash(self, file, algo):
-        hash = algo()
-        f = self.open_file(file)
-        while chunk := f.read(65536):
-            hash.update(chunk)
-        return hash
 
     def open_file(self, file):
         path = self.get_file_path(file)
