+++ common/iso_path_reader/methods/pycdlib.py
@@ -3,10 +3,11 @@ from pycdlib.pycdlib import _yield_children
 from pycdlib.pycdlibexception import PyCdlibInvalidInput
 
 from common.iso_path_reader.methods.base import IsoPathReader
+from common.iso_path_reader.methods.chunked_hash_trait import ChunkedHashTrait
 from dates import datetime_from_iso_date
 
 
-class PyCdLibPathReader(IsoPathReader):
+class PyCdLibPathReader(ChunkedHashTrait, IsoPathReader):
     def __init__(self, iso, fp, udf=False):
         super().__init__(iso, fp)
         self.udf = udf
@@ -74,14 +75,6 @@ class PyCdLibPathReader(IsoPathReader):
 
     def open_file(self, file):
         return pycdlib.pycdlibio.PyCdlibIO(file.inode, self.iso.logical_block_size)
-
-    def get_file_hash(self, file, algo):
-        hash = algo()
-        with self.open_file(file) as f:
-            for chunk in iter(lambda: f.read(65536), b""):
-                hash.update(chunk)
-
-        return hash
 
     def get_pvd(self):
         return self.iso.pvd
